1Hz telemetry data csv file converted from tlog
This script allows you to convert .tlog files generated by Ardusub during an ROV survey to .csv files. The data is clipped based on inputted start and end time for a survey. DVLlat and DVLlon coordinates are added along with the approximate area captured by the downward facing GoPro. Finally, a map is created displaying the ROV tracks based on the GPS and DVL coordinates.